Region and Platform IDs for game are always available
If the Regions and Platforms are embedded into the game, their ID
list is not available in the API. But we can simply create that
array out of the embedded regions and platforms.

Unfortunately this is not possible for the moderators as far as I
know, as the moderator type won't be available anymore.