Implemented optimized Game Non-Embed Calls
For a game, there are some lists of ids that can be quite large.
When they are not embedded, you'd have to query the API for every
single one of them, due to there not being a specialized API Call
for those. This can take quite a long time. Therefore it's better
to just query the game again with those embedded in these
situations. That reduces multiple API Calls into a single API, with
a slight overhead of having to parse the game again, which is okay,
since this is way faster than doing more than one API Call. If only
a single API Call is needed anyway, that API Call is done instead,
due to the overhead of embedding the result in a game.