Implemented Series
Series are now distinguished from games. There's also Rom Hacks now
that are basically games with the "IsRomHack" flag set as true.

Also querying API Elements from Site Uris is not done through
modifying the URL anymore, but instead queries the Uri and checks
for a Link Header that contains an API Uri. This is way safer.