Merged Variable Values and Choices into 1 class
Instead of using 2 classes for representing Variable Values, they
are a single class now that either contains 2 IDs or the Variable
and they Value ID + Name. All values that are not filled out based
on the case, are automatically dynamically loaded. Merging this
into a single class allows for easier filtering of leaderboard
variables, as you don't need to manually construct a Variable Value
anymore.